move search usecase tests to dedicated package
